+++ src/components/TableNode/TableNode.module.css
@@ -1,0 +1,41 @@
+.tableNode {
+  width: var(--baseWidthTableNode);
+  height: var(--baseHeightTableNode);
+  background-color: white;
+  border: 1.5px solid var(--borderColorTableNode);
+  border-radius: 2px;
+  display: flex;
+  flex-direction: column;
+  align-items: top;
+  justify-content: flex-start;
+  color: black;
+}
+
+.tableInput {
+  width: var(--baseWidthTableNode);
+  height: 50px;
+  border-radius: 2px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+}
+
+.tableInput:focus {
+  outline: none;
+  border: 0;
+  border-bottom: 1.5px solid var(--borderColorTableNode);
+  background-color: transparent;
+}
+
+.tableTitle {
+  width: var(--baseWidthTableNode);
+  height: 50px;
+  border-bottom: 1.5px solid var(--borderColorTableNode);
+  border-radius: 2px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+
+}
