--- src/styles/colors.css
@@ -1,3 +1,0 @@
-:root {
-  --whiteColor: #fff;
-}
