+++ src/styles/globals.css
@@ -1,4 +1,4 @@
-@import url('./colors.css');
+@import url('./variables.css');
 
 /* :root {
   --max-width: 1100px;
